Taichung Bike2.0 Park/requirements.txt
altgraph==0.17.3
asgiref==3.5.2
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
autopep8==1.7.0
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver-autoinstaller==0.4.0
colorama==0.4.5
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
Django==4.1.4
djangorestframework==3.14.0
entrypoints==0.4
exceptiongroup==1.0.4
executing==1.1.1
fonttools==4.38.0
future==0.18.2
h11==0.14.0
idna==3.4
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.4.2
kiwisolver==1.4.4
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.4
outcome==1.2.0
packaging==21.3
pandas==1.5.2
parso==0.8.3
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.13.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pywin32==304
pywin32-ctypes==0.2.0
pyzmq==24.0.1
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.1
toml==0.10.2
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
urllib3==1.26.13
wcwidth==0.2.5
webdriver-manager==3.8.5
wsproto==1.2.0
